Initial Changes After Backend Authentication Implemented

+++ src/apiCalls/deleteLike.ts
@@ -1,16 +1,29 @@
+"use server"
+
+import getAccessToken from "@/utils/getAccessTokenCookie";
+
 export default async function deleteLike(postId: string, userId: string) {
     try {
-    const res = await fetch(`${process.env.NEXT_PUBLIC_API_URL}/api/v1/likes`, {
-        method: "DELETE",
-        headers: {
-            "Content-Type": "application/json",
-        },
-        body: JSON.stringify({postId, userId}),
-    });
+
+        const accessToken = await getAccessToken()
+
+        if (!accessToken) {
+            console.warn("Attempted to add comment without an access token. User might not be logged in.");
+            return { success: false, data: false, message: "Error: You must be logged in to add a comment." };
+        }
+
+        const res = await fetch(`${process.env.NEXT_PUBLIC_API_URL}/api/v1/likes`, {
+            method: "DELETE",
+            headers: {
+                "Content-Type": "application/json",
+                "Authorization": `Bearer ${accessToken}`
+            },
+            body: JSON.stringify({ postId, userId }),
+        });
         const data = await res.json();
-        return {success: true, data: data};
+        return { success: true, data: data, message: "" };
     } catch (error) {
         console.error(error);
-        return {success: false, data: false};
+        return { success: false, data: false, message: "" };
     }
 }
